+++ src/test/java/stepDefs/Hook.java
@@ -1,6 +1,5 @@
 package stepDefs;
 
-import baseEntities.BaseApiTest;
 import baseEntities.BaseTest;
 import dataHelper.DataHelper;
 import factory.BrowserFactory;
@@ -11,7 +10,6 @@ import io.cucumber.java.Scenario;
 import io.qameta.allure.Step;
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
-import org.apache.http.HttpHeaders;
 import org.apache.http.protocol.HTTP;
 import org.openqa.selenium.OutputType;
 import org.openqa.selenium.TakesScreenshot;
@@ -26,29 +24,28 @@ public class Hook extends BaseTest {
         this.baseTest = baseTest;
     }
 
-    @BeforeAll
+    @BeforeAll(order = 10001)
     public static void prepareData() {
         expectedProject.setName(DataHelper.getAddProject().getName());
         expectedProject.setSummary(DataHelper.getAddProject().getSummary());
     }
 
-//    @BeforeAll
-//    public void setupApi(){
-//        RestAssured.baseURI = ReadProperties.getUrl();
-//        RestAssured.requestSpecification = given()
-//                .auth().oauth2(ReadProperties.token())
-//                .header(HTTP.CONTENT_TYPE, ContentType.JSON);
-//    }
-
-    @Before
+    @Before(value = "@gui",order = 10002)
     @Step("Browser initialization")
     public void initGUIScenario(Scenario scenario) {
         baseTest.driver = new BrowserFactory().getDriver();
-        baseTest.setupApi();
+    }
+
+    @Before(value = "@api")
+    public void setupApi(){
+        RestAssured.baseURI = ReadProperties.getUrl();
+        RestAssured.requestSpecification = given()
+                .auth().oauth2(ReadProperties.token())
+                .header(HTTP.CONTENT_TYPE, ContentType.JSON);
     }
 
 
-    @After
+    @After(value = "@gui")
     public void tearDown(Scenario scenario) {
         try {
             String screenshotName = scenario.getName().replaceAll("", "_");
