+++ src/test/resources/features/api/GetSingleProject.feature
@@ -1,6 +1,7 @@
+@api
 Feature: Get Single Project
 
   Scenario: Authorized user get particular project
-    Given GET project by id
+    Given GET project by projectId "60"
     Then status code is validated
     Then received ProjectId is validated
