+++ src/test/resources/features/gui/CreateEntity.feature
@@ -1,3 +1,4 @@
+@gui
 Feature: Create Project entity
 
   Scenario: Positive flow: Create Project entity
