+++ src/test/resources/features/gui/tooltip.feature
@@ -1,3 +1,4 @@
+@gui
 Feature: Display a pop-up message
 
   Scenario: Display a tooltip
